Skip Email
Skipped sending email, can refer to.
It is possible to reset the password in the personal profile, but the logic of forgetting the password has been modified and emails will no longer be sent to the user. Each startup can only be modified once; If you want to change your password again, please restart the backend.